requirements.txt
afinn==0.1
appnope==0.1.0
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
future==0.16.0
idna==2.6
invoke==0.21.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
matplotlib==2.1.0
numpy==1.13.3
oauthlib==2.0.6
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.2.3
python-dateutil==2.6.1
python-twitter==3.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
vaderSentiment==2.5
wcwidth==0.1.7
